+++ hass-actronque/Service.cs
@@ -29,7 +29,6 @@ namespace HMX.HASSActronQue
 			get { return _strDeviceNameMQTT; }
 		}
 
-
 		public static void Start()
         {
 			IConfigurationRoot configuration;
@@ -37,7 +36,7 @@ namespace HMX.HASSActronQue
 			string strMQTTUser, strMQTTPassword, strMQTTBroker;
 			string strQueUser, strQuePassword, strQueSerial, strSystemType;
 			int iPollInterval;
-			bool bPerZoneControls, bPerZoneSensors, bQueLogging, bMQTTLogging, bMQTTTLS, bDisableEventUpdates, bSeparateHeatCool;
+			bool bPerZoneControls, bQueLogging, bMQTTLogging, bMQTTTLS, bDisableEventUpdates, bSeparateHeatCool;
 
 			Logging.WriteDebugLog("Service.Start() Build Date: {0}", Properties.Resources.BuildDate);
 
@@ -68,8 +67,6 @@ namespace HMX.HASSActronQue
 
 			if (!Configuration.GetConfiguration(configuration, "PerZoneControls", out bPerZoneControls))
 				return;
-
-			Configuration.GetOptionalConfiguration(configuration, "PerZoneSensors", out bPerZoneSensors);		
 
 			if (!Configuration.GetConfiguration(configuration, "PollInterval", out iPollInterval) || iPollInterval < 10 || iPollInterval > 300)
 			{
@@ -118,7 +115,7 @@ namespace HMX.HASSActronQue
 
 			MQTT.StartMQTT(strMQTTBroker, bMQTTLogging, bMQTTTLS, _strServiceName, strMQTTUser, strMQTTPassword, MQTTProcessor);
 
-			Que.Initialise(strQueUser, strQuePassword, strQueSerial, strSystemType, iPollInterval, bQueLogging, bPerZoneControls, bPerZoneSensors, bDisableEventUpdates, bSeparateHeatCool, _eventStop);
+			Que.Initialise(strQueUser, strQuePassword, strQueSerial, strSystemType, iPollInterval, bQueLogging, bPerZoneControls, bDisableEventUpdates, bSeparateHeatCool, _eventStop);
 
 			webHost.Run();
 		}